refactor(cli): extract config bootstrap controller
